requirements.txt
fastapi~=0.104.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
python-multipart==0.0.6
passlib==1.7.4
python-jose
rich
python-multipart
psycopg2-binary==2.9.9